* remove single-threaded webcam demo * add inference fps control * add auto buffer-size * fix bugs in imshow_bbox
